refactor vimwiki
Taskwiki was slow and sluggish, there was the need for a speedy, simple
and reliable solution for note taking. Therefore I changed to
simplenotes. This allows for easy note taking (and viewing on multiple
devices).

I also switched the vimwiki syntax to vimwiki, to ahve the full
functionality. Vimwiki is now primarily used to cataloge my knowledge,
while simplenotes is for pure notetaking