Adding sncli
taskwarrior is nice to organize tasks, but i rather need to organize
notes. SImplenotes lets me easily sync accross devices, has a nice cli
which uses vim and simple!

The usage of the cli (sncli) is added. To keep the passwords from
pdot.conf a new file pdot_secret is created, which is not sourced.